+++ driver-bitmain.c
@@ -1860,7 +1860,7 @@ static struct cgpu_info * bitmain_usb_detect_one(libusb_device *dev, struct usb_
 	info->fan_pwm = BITMAIN_DEFAULT_FAN_MIN_PWM;
 	info->temp_max = 0;
 	/* This is for check the temp/fan every 3~4s */
-	info->temp_history_count = (4 / (float)((float)info->timeout * ((float)1.67/0x32))) + 1;
+	info->temp_history_count = (4 / ((float)info->timeout * ((float)1.67 / 0x32))) + 1;
 	if (info->temp_history_count <= 0)
 		info->temp_history_count = 1;
 
@@ -1961,7 +1961,7 @@ static bool bitmain_detect_one(const char * devpath)
 	info->fan_pwm = BITMAIN_DEFAULT_FAN_MIN_PWM;
 	info->temp_max = 0;
 	/* This is for check the temp/fan every 3~4s */
-	info->temp_history_count = (4 / (float)((float)info->timeout * ((float)1.67/0x32))) + 1;
+	info->temp_history_count = (4 / ((float)info->timeout * ((float)1.67 / 0x32))) + 1;
 	if (info->temp_history_count <= 0)
 		info->temp_history_count = 1;
